Use int type instead uint8_t

+++ hanoi/main.cpp
@@ -3,15 +3,14 @@
 #include <cstdlib>
 #include <algorithm>
 #include <string>
-#include <cstdint>
 #include <cassert>
 
 class Hanoi
 {
   public:
-    Hanoi(uint8_t disks) : disks(disks), steps(0)
+    Hanoi(int disks) : disks(disks), steps(0)
     {
-        for (uint8_t disk = disks; disk > 0; --disk)
+        for (int disk = disks; disk > 0; --disk)
         {
             pins[0].push_back(disk);
         }
@@ -20,19 +19,19 @@ class Hanoi
 
     void render()
     {
-        for (uint8_t height = disks; height > 0; --height)
+        for (int height = disks; height > 0; --height)
         {
-            for (uint8_t pin = 0; pin < 3; ++pin)
+            for (int pin = 0; pin < 3; ++pin)
             {
                 if (pins[pin].size() >= height)
                 {
-                    uint8_t dw      = diskWidth(pins[pin][height - 1]);
-                    uint8_t padding = (maxDiskWidth - dw) / 2;
+                    int dw      = diskWidth(pins[pin][height - 1]);
+                    int padding = (maxDiskWidth - dw) / 2;
                     std::cout << std::string(padding, ' ') + std::string(dw, '=') + std::string(padding, ' ') + ' ';
                 }
                 else
                 {
-                    uint8_t padding = (maxDiskWidth - 1) / 2;
+                    int padding = (maxDiskWidth - 1) / 2;
                     std::cout << std::string(padding, ' ') + '|' + std::string(padding, ' ') + ' ';
                 }
             }
@@ -45,17 +44,17 @@ class Hanoi
     {
         render();
         moveTower(0, 2, disks);
-        std::cout << "Algorithm used " << (int)steps << " steps" << std::endl;
+        std::cout << "Algorithm used " << steps << " steps" << std::endl;
     }
 
   private:
-    uint8_t disks;
-    uint8_t maxDiskWidth;
-    std::vector<uint8_t> pins[3];
-    uint8_t steps;
+    int disks;
+    int maxDiskWidth;
+    std::vector<int> pins[3];
+    int steps;
 
-    uint8_t diskWidth(uint8_t disk) { return 1 + 2 * (disk - 1); }
-    uint8_t findFreePin(uint8_t pin1, uint8_t pin2)
+    int diskWidth(int disk) { return 1 + 2 * (disk - 1); }
+    int findFreePin(int pin1, int pin2)
     {
         assert(0 <= pin1 && pin1 <= 2);
         assert(0 <= pin2 && pin2 <= 2);
@@ -63,12 +62,12 @@ class Hanoi
         return std::abs(pin1 + pin2 - 3);
     }
 
-    void moveDisk(uint8_t fromPin, uint8_t toPin)
+    void moveDisk(int fromPin, int toPin)
     {
         assert(0 <= fromPin && fromPin <= 2);
         assert(0 <= toPin && toPin <= 2);
 
-        uint8_t disk = pins[fromPin].back();
+        int disk = pins[fromPin].back();
         pins[fromPin].pop_back();
         pins[toPin].push_back(disk);
 
@@ -77,13 +76,13 @@ class Hanoi
         render();
     }
 
-    void moveTower(uint8_t fromPin, uint8_t toPin, uint8_t size)
+    void moveTower(int fromPin, int toPin, int size)
     {
         assert(0 <= fromPin && fromPin <= 2);
         assert(0 <= toPin && toPin <= 2);
         assert(0 <= size && size <= disks);
 
-        uint8_t freePin = findFreePin(fromPin, toPin);
+        int freePin = findFreePin(fromPin, toPin);
 
         if (size == 0)
             return;
@@ -104,7 +103,7 @@ int main(int argc, char** argv)
         return -1;
     }
 
-    uint8_t disk_number = std::stoul(std::string(argv[1]));
+    int disk_number = std::stoul(std::string(argv[1]));
     Hanoi hanoi(disk_number);
     hanoi.run();
